# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChordViewer local API
  version: 1.3.0
  description: >-
    Workstation API. Browser authentication uses HttpOnly cookies and a trusted Origin;
    native clients use signed bearer sessions from set-auth-token. Public deployment is deferred.
    Sheet writes are limited to 1 MiB and libraries to 100 sheets per account,
    including recoverable Trash. Reads never change recent-open order.
servers:
  - url: http://127.0.0.1:3000
paths:
  /health:
    get:
      operationId: getHealth
      summary: Report local API availability
      responses:
        '200':
          description: API and migrated database are ready. Unavailable database returns 503.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, service, apiVersion, database]
                properties:
                  status: { type: string, const: ok }
                  service: { type: string, const: chordviewer-api }
                  apiVersion: { type: string, const: '1' }
                  database: { type: string, const: ready }
  /api/v1/score-example:
    get:
      operationId: getScoreExample
      summary: Return the original shared four-measure notation proof
      responses:
        '200':
          description: The exact shared score fixture, validated before serving.
          content:
            application/json:
              schema:
                $ref: './src/lead-sheet-v1.schema.json'
  /api/auth/sign-up/email:
    post:
      summary: Create a local email/password account and sign in
      description: >-
        Body contains name (1–100 characters), email (up to 254), and password (12–128).
        Sets the browser cookie. Requests without Origin receive the signed set-auth-token
        header for native use. JSON never contains a raw token. Email verification/recovery
        and external authentication are not enabled in this private milestone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, email, password]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
      responses:
        '200': { description: Account created and session established. }
        '400': { description: Invalid input or registration rejected. }
        '403': { description: Invalid request origin or Fetch Metadata. }
        '429': { description: Authentication rate limit reached. }
  /api/auth/sign-in/email:
    post:
      summary: Sign in to an existing local account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
      responses:
        '200': { description: Session established; same cookie/native-header behavior as registration. }
        '401': { description: Sign-in failed. }
        '429': { description: Authentication rate limit reached. }
  /api/auth/sign-out:
    post:
      summary: Revoke the current session
      security: [{ browserSession: [] }, { nativeSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        '200': { description: Current session revoked and browser cookie cleared. }
  /api/v1/me:
    get:
      summary: Return the authenticated account's id, name and email
      security: [{ browserSession: [] }, { nativeSession: [] }]
      responses:
        '200': { description: 'Object with user: {id, name, email}; no session credential.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/sheets:
    get:
      summary: List this account's active and trashed sheet summaries
      security: [{ browserSession: [] }, { nativeSession: [] }]
      responses:
        '200':
          description: Bounded metadata and at most four first-bar chord labels per sheet; a new account receives an empty array.
          content:
            application/json:
              schema:
                type: object
                required: [sheets]
                properties:
                  sheets:
                    type: array
                    maxItems: 100
                    items: { $ref: '#/components/schemas/SheetSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create a blank sheet or an explicit copy of the original example
      description: >-
        Blank sheets use score version 2 with the requested key and meter (defaults C major, 4/4).
        Example copies preserve the original C-major 4/4 notation; other settings are rejected.
      security: [{ browserSession: [] }, { nativeSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [title, template]
              properties:
                title: { type: string, minLength: 1, maxLength: 200 }
                template: { type: string, enum: [blank, example] }
                tutorialUrl: { $ref: '#/components/schemas/Tutorial' }
                keySignature: { $ref: './src/lead-sheet-v2.schema.json#/properties/keySignature' }
                timeSignature: { $ref: './src/lead-sheet-v2.schema.json#/properties/timeSignature' }
      responses:
        '201': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid title, template, tutorial, key, meter or extra fields. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: 'sheet_limit: account already has 100 sheets, including Trash.' }
        '413': { description: Request exceeds 1 MiB. }
  /api/v1/sheets/import:
    post:
      summary: Save a validated imported score as a new sheet
      description: >-
        The client previews and converts its local file before this request. The backend accepts
        only supported score JSON, validates it again, replaces the source score ID with a fresh
        server ID and uses the supplied title. It never updates a source sheet or fetches a URL.
        Creation is atomic and shares the normal 100-sheet account limit. Ownership comes only
        from the session. Browser cookie writes require a trusted Origin. Do not automatically
        retry an import when the network outcome is unknown, because that may create another copy.
      security: [{ browserSession: [] }, { nativeSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [score, title]
              properties:
                score: { $ref: '#/components/schemas/Score' }
                title: { type: string, minLength: 1, maxLength: 200 }
                tutorialUrl: { $ref: '#/components/schemas/Tutorial' }
      responses:
        '201': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid score, title, tutorial or extra fields. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '409': { description: 'sheet_limit: account already has 100 sheets, including Trash.' }
        '413': { description: Request exceeds 1 MiB including its JSON envelope. }
  /api/v1/sheets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Read an active sheet owned by the authenticated account without changing recency
      security: [{ browserSession: [] }, { nativeSession: [] }]
      responses:
        '200': { $ref: '#/components/responses/Saved' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Unknown sheet or sheet belongs to another account. }
    put:
      summary: Save a complete validated score and tutorial against its current revision
      description: >-
        The score ID is immutable. Updates compare expectedRevision atomically, then increment it.
        Browser cookie writes require a trusted Origin; native writes require a signed bearer token.
        Successful responses reflect the authoritative saved state. Clients must not retry creates
        automatically when the network outcome is unknown.
      security: [{ browserSession: [] }, { nativeSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [score, tutorialUrl, expectedRevision]
              properties:
                score: { $ref: '#/components/schemas/Score' }
                tutorialUrl: { $ref: '#/components/schemas/Tutorial' }
                expectedRevision: { type: integer, minimum: 1, maximum: 2147483646 }
      responses:
        '200': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid score, identity, revision or tutorial. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '404': { description: Unknown sheet or sheet belongs to another account. }
        '409': { description: 'revision_conflict: reload before another save.' }
        '413': { description: Request exceeds 1 MiB. }
  /api/v1/sheets/{id}/open:
    post:
      summary: Record an intentional user open for Recent sorting
      security: [{ browserSession: [] }, { nativeSession: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: false } } }
      responses:
        '200': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '404': { description: Unknown, trashed or unowned sheet. }
  /api/v1/sheets/{id}/metadata:
    patch:
      summary: Rename, favorite or designate a Draft using an atomic revision check
      security: [{ browserSession: [] }, { nativeSession: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision]
              anyOf:
                - required: [title]
                - required: [favorite]
                - required: [draft]
              properties:
                expectedRevision: { type: integer, minimum: 1, maximum: 2147483646 }
                title: { type: string, minLength: 1, maxLength: 200 }
                favorite: { type: boolean }
                draft: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid metadata or revision. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '404': { description: Unknown, trashed or unowned sheet. }
        '409': { description: Revision conflict. }
  /api/v1/sheets/{id}/duplicate:
    post:
      summary: Create a fresh-ID copy of an active sheet
      security: [{ browserSession: [] }, { nativeSession: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RevisionRequest' } } }
      responses:
        '201': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid revision. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '404': { description: Unknown, trashed or unowned source sheet. }
        '409': { description: Revision conflict or 100-sheet limit including Trash. }
  /api/v1/sheets/{id}/trash:
    post:
      summary: Move an active sheet into recoverable Trash
      security: [{ browserSession: [] }, { nativeSession: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RevisionRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid revision. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '404': { description: Unknown or unowned sheet. }
        '409': { description: Revision conflict. }
  /api/v1/sheets/{id}/restore:
    post:
      summary: Restore a sheet from Trash
      security: [{ browserSession: [] }, { nativeSession: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RevisionRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/Saved' }
        '400': { description: Invalid revision. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Missing or untrusted browser Origin. }
        '404': { description: Unknown or unowned sheet. }
        '409': { description: Revision conflict. }
components:
  securitySchemes:
    browserSession: { type: apiKey, in: cookie, name: chordviewer.session_token }
    nativeSession: { type: http, scheme: bearer, bearerFormat: Signed opaque database session }
  responses:
    Unauthorized:
      description: Missing, invalid, expired or revoked session.
    Saved:
      description: Persisted sheet and current revision.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SavedSheet' }
  schemas:
    RevisionRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties:
        expectedRevision: { type: integer, minimum: 1, maximum: 2147483646 }
    Score:
      description: Version 1 remains readable; version 2 adds supported global keys and meters.
      oneOf:
        - { $ref: './src/lead-sheet-v1.schema.json' }
        - { $ref: './src/lead-sheet-v2.schema.json' }
    Tutorial:
      type: [string, 'null']
      maxLength: 500
      description: >-
        Null for no video; HTTPS youtube.com, www.youtube.com, m.youtube.com watch?v=ID
        or youtu.be/ID with an eleven-character video ID. Stored in canonical www.youtube.com/watch form.
        No backend URL fetching occurs.
    SheetMetadata:
      type: object
      required: [id, tutorialUrl, revision, createdAt, updatedAt, favorite, draft, trashedAt, openedAt]
      properties:
        id: { type: string, format: uuid }
        tutorialUrl: { $ref: '#/components/schemas/Tutorial' }
        revision: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        favorite: { type: boolean }
        draft: { type: boolean }
        trashedAt: { type: [string, 'null'], format: date-time }
        openedAt: { type: [string, 'null'], format: date-time }
    SheetSummary:
      allOf:
        - { $ref: '#/components/schemas/SheetMetadata' }
        - type: object
          required: [title, keySignature, timeSignature, hasChords, hasMelody, previewChords]
          properties:
            title: { type: string, minLength: 1, maxLength: 200 }
            keySignature: { $ref: './src/lead-sheet-v2.schema.json#/properties/keySignature' }
            timeSignature: { $ref: './src/lead-sheet-v2.schema.json#/properties/timeSignature' }
            hasChords: { type: boolean }
            hasMelody: { type: boolean }
            previewChords: { type: array, maxItems: 4, items: { type: string, maxLength: 64 } }
    SavedSheet:
      allOf:
        - { $ref: '#/components/schemas/SheetMetadata' }
        - type: object
          required: [score]
          properties:
            score: { $ref: '#/components/schemas/Score' }